Refactoring into AbstractSpout for ES
