+++ external/elasticsearch/src/main/java/com/digitalpebble/stormcrawler/elasticsearch/persistence/AbstractSpout.java
@@ -1,0 +1,181 @@
+/**
+ * Licensed to DigitalPebble Ltd under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * DigitalPebble licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.digitalpebble.stormcrawler.elasticsearch.persistence;
+
+import java.util.Iterator;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Queue;
+
+import org.apache.storm.metric.api.IMetric;
+import org.apache.storm.metric.api.MultiCountMetric;
+import org.apache.storm.spout.SpoutOutputCollector;
+import org.apache.storm.task.TopologyContext;
+import org.apache.storm.topology.OutputFieldsDeclarer;
+import org.apache.storm.topology.base.BaseRichSpout;
+import org.apache.storm.tuple.Fields;
+import org.apache.storm.tuple.Values;
+import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsGroup;
+import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsRequest;
+import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsResponse;
+import org.elasticsearch.client.Client;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.digitalpebble.stormcrawler.Metadata;
+import com.digitalpebble.stormcrawler.elasticsearch.ElasticSearchConnection;
+import com.digitalpebble.stormcrawler.util.ConfUtils;
+
+public abstract class AbstractSpout extends BaseRichSpout {
+
+    private static final Logger LOG = LoggerFactory
+            .getLogger(AbstractSpout.class);
+
+    protected static final String ESBoltType = "status";
+    protected static final String ESStatusIndexNameParamName = "es.status.index.name";
+    protected static final String ESStatusDocTypeParamName = "es.status.doc.type";
+
+    protected String indexName;
+    protected String docType;
+    protected boolean active = true;
+    protected SpoutOutputCollector _collector;
+    protected MultiCountMetric eventCounter;
+
+    protected static Client client;
+
+    /**
+     * when using multiple instances - each one is in charge of a specific shard
+     * useful when sharding based on host or domain to guarantee a good mix of
+     * URLs
+     */
+    protected int shardID = -1;
+
+    /** Used to distinguish between instances in the logs **/
+    protected String logIdprefix = "";
+
+    protected Queue<Values> buffer = new LinkedList<>();
+
+    @Override
+    public void open(Map stormConf, TopologyContext context,
+            SpoutOutputCollector collector) {
+
+        indexName = ConfUtils.getString(stormConf, ESStatusIndexNameParamName,
+                "status");
+        docType = ConfUtils.getString(stormConf, ESStatusDocTypeParamName,
+                "status");
+
+        // one ES client per JVM
+        synchronized (AbstractSpout.class) {
+            try {
+                if (client == null) {
+                    client = ElasticSearchConnection.getClient(stormConf,
+                            ESBoltType);
+                }
+            } catch (Exception e1) {
+                LOG.error("Can't connect to ElasticSearch", e1);
+                throw new RuntimeException(e1);
+            }
+        }
+
+        // if more than one instance is used we expect their number to be the
+        // same as the number of shards
+        int totalTasks = context
+                .getComponentTasks(context.getThisComponentId()).size();
+        if (totalTasks > 1) {
+            logIdprefix = "[" + context.getThisComponentId() + " #"
+                    + context.getThisTaskIndex() + "] ";
+
+            // determine the number of shards so that we can restrict the
+            // search
+            ClusterSearchShardsRequest request = new ClusterSearchShardsRequest(
+                    indexName);
+            ClusterSearchShardsResponse shardresponse = client.admin()
+                    .cluster().searchShards(request).actionGet();
+            ClusterSearchShardsGroup[] shardgroups = shardresponse.getGroups();
+            if (totalTasks != shardgroups.length) {
+                throw new RuntimeException(
+                        "Number of ES spout instances should be the same as number of shards ("
+                                + shardgroups.length + ") but is " + totalTasks);
+            }
+            shardID = shardgroups[context.getThisTaskIndex()].getShardId();
+            LOG.info("{} assigned shard ID {}", logIdprefix, shardID);
+        }
+
+        _collector = collector;
+
+        eventCounter = context.registerMetric("counters",
+                new MultiCountMetric(), 10);
+
+        context.registerMetric("buffer_size", new IMetric() {
+            @Override
+            public Object getValueAndReset() {
+                return buffer.size();
+            }
+        }, 10);
+    }
+
+    @Override
+    public void declareOutputFields(OutputFieldsDeclarer declarer) {
+        declarer.declare(new Fields("url", "metadata"));
+    }
+
+    protected final Metadata fromKeyValues(Map<String, Object> keyValues) {
+        Map<String, List<String>> mdAsMap = (Map<String, List<String>>) keyValues
+                .get("metadata");
+        Metadata metadata = new Metadata();
+        if (mdAsMap != null) {
+            Iterator<Entry<String, List<String>>> mdIter = mdAsMap.entrySet()
+                    .iterator();
+            while (mdIter.hasNext()) {
+                Entry<String, List<String>> mdEntry = mdIter.next();
+                String key = mdEntry.getKey();
+                // periods are not allowed in ES2 - replace with %2E
+                key = key.replaceAll("%2E", "\\.");
+                Object mdValObj = mdEntry.getValue();
+                // single value
+                if (mdValObj instanceof String) {
+                    metadata.addValue(key, (String) mdValObj);
+                }
+                // multi valued
+                else {
+                    metadata.addValues(key, (List<String>) mdValObj);
+                }
+            }
+        }
+        return metadata;
+    }
+
+    @Override
+    public void activate() {
+        active = true;
+    }
+
+    @Override
+    public void deactivate() {
+        active = false;
+    }
+
+    @Override
+    public void close() {
+        if (client != null)
+            client.close();
+    }
+
+}
