+++ external/elasticsearch/src/main/java/com/digitalpebble/stormcrawler/elasticsearch/persistence/AggregationSpout.java
@@ -20,22 +20,20 @@ package com.digitalpebble.stormcrawler.elasticsearch.persistence;
 import java.util.Collections;
 import java.util.Date;
 import java.util.HashSet;
-import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Queue;
 import java.util.Set;
 
 import org.apache.commons.lang.StringUtils;
-import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsGroup;
-import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsRequest;
-import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsResponse;
+import org.apache.storm.metric.api.IMetric;
+import org.apache.storm.spout.SpoutOutputCollector;
+import org.apache.storm.task.TopologyContext;
+import org.apache.storm.tuple.Values;
 import org.elasticsearch.action.search.SearchRequestBuilder;
 import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.action.search.SearchType;
-import org.elasticsearch.client.Client;
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
 import org.elasticsearch.search.SearchHit;
@@ -53,17 +51,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.digitalpebble.stormcrawler.Metadata;
-import com.digitalpebble.stormcrawler.elasticsearch.ElasticSearchConnection;
 import com.digitalpebble.stormcrawler.util.ConfUtils;
-
-import org.apache.storm.metric.api.IMetric;
-import org.apache.storm.metric.api.MultiCountMetric;
-import org.apache.storm.spout.SpoutOutputCollector;
-import org.apache.storm.task.TopologyContext;
-import org.apache.storm.topology.OutputFieldsDeclarer;
-import org.apache.storm.topology.base.BaseRichSpout;
-import org.apache.storm.tuple.Fields;
-import org.apache.storm.tuple.Values;
 
 /**
  * Spout which pulls URL from an ES index. Use a single instance unless you use
@@ -73,18 +61,12 @@ import org.apache.storm.tuple.Values;
  * metadata.hostname.
  **/
 @SuppressWarnings("serial")
-public class AggregationSpout extends BaseRichSpout {
+public class AggregationSpout extends AbstractSpout {
 
     private static final Logger LOG = LoggerFactory
             .getLogger(AggregationSpout.class);
 
-    private static final String ESBoltType = "status";
-
-    private static final String ESStatusIndexNameParamName = "es.status.index.name";
-    private static final String ESStatusDocTypeParamName = "es.status.doc.type";
-
     private static final String ESStatusRoutingFieldParamName = "es.status.routing.fieldname";
-
     private static final String ESStatusMaxBucketParamName = "es.status.max.buckets";
     private static final String ESStatusMaxURLsParamName = "es.status.max.urls.per.bucket";
 
@@ -105,16 +87,7 @@ public class AggregationSpout extends BaseRichSpout {
      **/
     private static final String ESStatusMinDelayParamName = "es.status.min.delay.queries";
 
-    protected String indexName;
-    protected String docType;
-
-    private SpoutOutputCollector _collector;
-
-    protected static Client client;
-
     protected Set<String> beingProcessed = new HashSet<>();
-
-    protected Queue<Values> buffer = new LinkedList<>();
 
     /** Field name used for field collapsing e.g. metadata.hostname **/
     protected String partitionField;
@@ -123,18 +96,7 @@ public class AggregationSpout extends BaseRichSpout {
 
     protected int maxBucketNum = 10;
 
-    protected MultiCountMetric eventCounter;
-
-    private boolean active = true;
-
     protected long minDelayBetweenQueries = 2000;
-
-    /**
-     * when using multiple instances - each one is in charge of a specific shard
-     * useful when sharding based on host or domain to guarantee a good mix of
-     * URLs
-     */
-    protected int shardID = -1;
 
     private String bucketSortField = "";
 
@@ -142,17 +104,9 @@ public class AggregationSpout extends BaseRichSpout {
 
     protected Date timePreviousQuery = null;
 
-    /** Used to distinguish between instances in the logs **/
-    protected String logIdprefix = "";
-
     @Override
     public void open(Map stormConf, TopologyContext context,
             SpoutOutputCollector collector) {
-
-        indexName = ConfUtils.getString(stormConf, ESStatusIndexNameParamName,
-                "status");
-        docType = ConfUtils.getString(stormConf, ESStatusDocTypeParamName,
-                "status");
 
         partitionField = ConfUtils.getString(stormConf,
                 ESStatusRoutingFieldParamName);
@@ -171,65 +125,7 @@ public class AggregationSpout extends BaseRichSpout {
         minDelayBetweenQueries = ConfUtils.getLong(stormConf,
                 ESStatusMinDelayParamName, 2000);
 
-        // one ES client per JVM
-        synchronized (AggregationSpout.class) {
-            try {
-                if (client == null) {
-                    client = ElasticSearchConnection.getClient(stormConf,
-                            ESBoltType);
-                }
-            } catch (Exception e1) {
-                LOG.error("Can't connect to ElasticSearch", e1);
-                throw new RuntimeException(e1);
-            }
-        }
-
-        // if more than one instance is used we expect their number to be the
-        // same as the number of shards
-        int totalTasks = context
-                .getComponentTasks(context.getThisComponentId()).size();
-        if (totalTasks > 1) {
-            logIdprefix = "[" + context.getThisComponentId() + " #"
-                    + context.getThisTaskIndex() + "] ";
-
-            // determine the number of shards so that we can restrict the
-            // search
-            ClusterSearchShardsRequest request = new ClusterSearchShardsRequest(
-                    indexName);
-            ClusterSearchShardsResponse shardresponse = client.admin()
-                    .cluster().searchShards(request).actionGet();
-            ClusterSearchShardsGroup[] shardgroups = shardresponse.getGroups();
-            if (totalTasks != shardgroups.length) {
-                throw new RuntimeException(
-                        "Number of ES spout instances should be the same as number of shards ("
-                                + shardgroups.length + ") but is " + totalTasks);
-            }
-            shardID = shardgroups[context.getThisTaskIndex()].getShardId();
-            LOG.info("{} assigned shard ID {}", logIdprefix, shardID);
-        }
-
-        _collector = collector;
-
-        this.eventCounter = context.registerMetric("counters",
-                new MultiCountMetric(), 10);
-
-        context.registerMetric("buffer_size", new IMetric() {
-            @Override
-            public Object getValueAndReset() {
-                return buffer.size();
-            }
-        }, 10);
-    }
-
-    @Override
-    public void close() {
-        if (client != null)
-            client.close();
-    }
-
-    @Override
-    public void declareOutputFields(OutputFieldsDeclarer declarer) {
-        declarer.declare(new Fields("url", "metadata"));
+        super.open(stormConf, context, collector);
     }
 
     @Override
@@ -246,7 +142,7 @@ public class AggregationSpout extends BaseRichSpout {
             String url = fields.get(0).toString();
             beingProcessed.add(url);
 
-            this._collector.emit(fields, url);
+            _collector.emit(fields, url);
             eventCounter.scope("emitted").incrBy(1);
 
             return;
@@ -388,32 +284,6 @@ public class AggregationSpout extends BaseRichSpout {
         eventCounter.scope("ES_docs").incrBy(numhits);
     }
 
-    protected final Metadata fromKeyValues(Map<String, Object> keyValues) {
-        Map<String, List<String>> mdAsMap = (Map<String, List<String>>) keyValues
-                .get("metadata");
-        Metadata metadata = new Metadata();
-        if (mdAsMap != null) {
-            Iterator<Entry<String, List<String>>> mdIter = mdAsMap.entrySet()
-                    .iterator();
-            while (mdIter.hasNext()) {
-                Entry<String, List<String>> mdEntry = mdIter.next();
-                String key = mdEntry.getKey();
-                // periods are not allowed in ES2 - replace with %2E
-                key = key.replaceAll("%2E", "\\.");
-                Object mdValObj = mdEntry.getValue();
-                // single value
-                if (mdValObj instanceof String) {
-                    metadata.addValue(key, (String) mdValObj);
-                }
-                // multi valued
-                else {
-                    metadata.addValues(key, (List<String>) mdValObj);
-                }
-            }
-        }
-        return metadata;
-    }
-
     @Override
     public void ack(Object msgId) {
         LOG.debug("{}  Ack for {}", logIdprefix, msgId);
@@ -426,16 +296,6 @@ public class AggregationSpout extends BaseRichSpout {
         LOG.info("{}  Fail for {}", logIdprefix, msgId);
         beingProcessed.remove(msgId);
         eventCounter.scope("failed").incrBy(1);
-    }
-
-    @Override
-    public void activate() {
-        active = true;
-    }
-
-    @Override
-    public void deactivate() {
-        active = false;
     }
 
 }
