+++ external/elasticsearch/src/main/java/com/digitalpebble/stormcrawler/elasticsearch/persistence/ElasticSearchSpout.java
@@ -19,21 +19,16 @@ package com.digitalpebble.stormcrawler.elasticsearch.persistence;
 
 import java.util.Date;
 import java.util.HashMap;
-import java.util.Iterator;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Queue;
 import java.util.concurrent.atomic.AtomicInteger;
 
-import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsGroup;
-import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsRequest;
-import org.elasticsearch.action.admin.cluster.shards.ClusterSearchShardsResponse;
+import org.apache.storm.metric.api.IMetric;
+import org.apache.storm.spout.SpoutOutputCollector;
+import org.apache.storm.task.TopologyContext;
+import org.apache.storm.tuple.Values;
 import org.elasticsearch.action.search.SearchRequestBuilder;
 import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.action.search.SearchType;
-import org.elasticsearch.client.Client;
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
 import org.elasticsearch.index.query.functionscore.FunctionScoreQueryBuilder;
@@ -46,49 +41,26 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.digitalpebble.stormcrawler.Metadata;
-import com.digitalpebble.stormcrawler.elasticsearch.ElasticSearchConnection;
 import com.digitalpebble.stormcrawler.util.ConfUtils;
 import com.digitalpebble.stormcrawler.util.URLPartitioner;
-
-import org.apache.storm.metric.api.IMetric;
-import org.apache.storm.metric.api.MultiCountMetric;
-import org.apache.storm.spout.SpoutOutputCollector;
-import org.apache.storm.task.TopologyContext;
-import org.apache.storm.topology.OutputFieldsDeclarer;
-import org.apache.storm.topology.base.BaseRichSpout;
-import org.apache.storm.tuple.Fields;
-import org.apache.storm.tuple.Values;
 
 /**
  * Spout which pulls URL from an ES index. Use a single instance unless you use
  * 'es.status.routing' with the StatusUpdaterBolt, in which case you need to
  * have exactly the same number of spout instances as ES shards.
  **/
-public class ElasticSearchSpout extends BaseRichSpout {
+public class ElasticSearchSpout extends AbstractSpout {
 
     private static final Logger LOG = LoggerFactory
             .getLogger(ElasticSearchSpout.class);
 
-    private static final String ESBoltType = "status";
-
-    private static final String ESStatusIndexNameParamName = "es.status.index.name";
-    private static final String ESStatusDocTypeParamName = "es.status.doc.type";
     private static final String ESStatusBufferSizeParamName = "es.status.max.buffer.size";
     private static final String ESStatusMaxInflightParamName = "es.status.max.inflight.urls.per.bucket";
     private static final String ESRandomSortParamName = "es.status.random.sort";
     private static final String ESMaxSecsSinceQueriedDateParamName = "es.status.max.secs.date";
     private static final String ESStatusSortFieldParamName = "es.status.sort.field";
 
-    private String indexName;
-    private String docType;
-
-    private SpoutOutputCollector _collector;
-
-    private static Client client;
-
     private int maxBufferSize = 100;
-
-    private Queue<Values> buffer = new LinkedList<>();
 
     private int lastStartOffset = 0;
     private Date lastDate;
@@ -109,10 +81,6 @@ public class ElasticSearchSpout extends BaseRichSpout {
     // URL / politeness bucket (hostname / domain etc...)
     private Map<String, String> beingProcessed = new HashMap<>();
 
-    private MultiCountMetric eventCounter;
-
-    private boolean active = true;
-
     // when using multiple instances - each one is in charge of a specific shard
     // useful when sharding based on host or domain to guarantee a good mix of
     // URLs
@@ -124,10 +92,6 @@ public class ElasticSearchSpout extends BaseRichSpout {
     public void open(Map stormConf, TopologyContext context,
             SpoutOutputCollector collector) {
 
-        indexName = ConfUtils.getString(stormConf, ESStatusIndexNameParamName,
-                "status");
-        docType = ConfUtils.getString(stormConf, ESStatusDocTypeParamName,
-                "status");
         maxInFlightURLsPerBucket = ConfUtils.getInt(stormConf,
                 ESStatusMaxInflightParamName, 1);
         maxBufferSize = ConfUtils.getInt(stormConf,
@@ -140,47 +104,10 @@ public class ElasticSearchSpout extends BaseRichSpout {
         sortField = ConfUtils.getString(stormConf, ESStatusSortFieldParamName,
                 "nextFetchDate");
 
-        // one ES client per JVM
-        synchronized (ElasticSearchSpout.class) {
-            try {
-                if (client == null) {
-                    client = ElasticSearchConnection.getClient(stormConf,
-                            ESBoltType);
-                }
-            } catch (Exception e1) {
-                LOG.error("Can't connect to ElasticSearch", e1);
-                throw new RuntimeException(e1);
-            }
-        }
-
-        // if more than one instance is used we expect their number to be the
-        // same as the number of shards
-        int totalTasks = context
-                .getComponentTasks(context.getThisComponentId()).size();
-        if (totalTasks > 1) {
-            // determine the number of shards so that we can restrict the
-            // search
-            ClusterSearchShardsRequest request = new ClusterSearchShardsRequest(
-                    indexName);
-            ClusterSearchShardsResponse shardresponse = client.admin()
-                    .cluster().searchShards(request).actionGet();
-            ClusterSearchShardsGroup[] shardgroups = shardresponse.getGroups();
-            if (totalTasks != shardgroups.length) {
-                throw new RuntimeException(
-                        "Number of ES spout instances should be the same as number of shards ("
-                                + shardgroups.length + ") but is " + totalTasks);
-            }
-            shardID = shardgroups[context.getThisTaskIndex()].getShardId();
-            LOG.info("Assigned shard ID {}", shardID);
-        }
+        super.open(stormConf, context, collector);
 
         partitioner = new URLPartitioner();
         partitioner.configure(stormConf);
-
-        _collector = collector;
-
-        this.eventCounter = context.registerMetric("counters",
-                new MultiCountMetric(), 10);
 
         context.registerMetric("beingProcessed", new IMetric() {
             @Override
@@ -188,24 +115,6 @@ public class ElasticSearchSpout extends BaseRichSpout {
                 return beingProcessed.size();
             }
         }, 10);
-
-        context.registerMetric("buffer_size", new IMetric() {
-            @Override
-            public Object getValueAndReset() {
-                return buffer.size();
-            }
-        }, 10);
-    }
-
-    @Override
-    public void close() {
-        if (client != null)
-            client.close();
-    }
-
-    @Override
-    public void declareOutputFields(OutputFieldsDeclarer declarer) {
-        declarer.declare(new Fields("url", "metadata"));
     }
 
     @Override
@@ -343,28 +252,7 @@ public class ElasticSearchSpout extends BaseRichSpout {
                 continue;
             }
 
-            Map<String, List<String>> mdAsMap = (Map<String, List<String>>) keyValues
-                    .get("metadata");
-            Metadata metadata = new Metadata();
-            if (mdAsMap != null) {
-                Iterator<Entry<String, List<String>>> mdIter = mdAsMap
-                        .entrySet().iterator();
-                while (mdIter.hasNext()) {
-                    Entry<String, List<String>> mdEntry = mdIter.next();
-                    String key = mdEntry.getKey();
-                    // periods are not allowed in ES2 - replace with %2E
-                    key = key.replaceAll("%2E", "\\.");
-                    Object mdValObj = mdEntry.getValue();
-                    // single value
-                    if (mdValObj instanceof String) {
-                        metadata.addValue(key, (String) mdValObj);
-                    }
-                    // multi valued
-                    else {
-                        metadata.addValues(key, (List<String>) mdValObj);
-                    }
-                }
-            }
+            Metadata metadata = fromKeyValues(keyValues);
             buffer.add(new Values(url, metadata));
         }
     }
@@ -393,16 +281,6 @@ public class ElasticSearchSpout extends BaseRichSpout {
         int newVal = currentValue.decrementAndGet();
         if (newVal == 0)
             this.inFlightTracker.remove(partitionKey);
-    }
-
-    @Override
-    public void activate() {
-        active = true;
-    }
-
-    @Override
-    public void deactivate() {
-        active = false;
     }
 
 }
